==> 03_mcnulty/notebooks/03_PWS_Cleaning.ipynb <==
import pandas as pd
import numpy as np
from dbfread import DBF
import pickle

def load_dbf_to_df(file_name):
    dbf = DBF(data_path + file_name, load=True, raw=True)
    df = pd.DataFrame(dbf.records)
    return df.applymap(lambda x: x.rstrip())

#pickle objects
def make_pickle(obj, filename):
    with open(filename, 'wb') as picklefile:
        pickle.dump(obj, picklefile)

#open pickled objects
def open_pickle(filename):
    with open(filename, 'rb') as picklefile:
        return pickle.load(picklefile)

### MassGIS Public Water Sources

data_path = '../data/external/MassGIS/rs101416/'

** PWSDEP_PT_WQTS **

The PWS_DEP.WQTS table contains information downloaded from the DWP Water Quality Testing System (WQTS) Oracle database. WQTS is DEP’s comprehensive repository for PWS information. The WQTS database contains many other data items that are not in the PWSDEP_PT_WQTS relate file. This table was designed for use by the DEP GIS Program to maintain the PWS datalayer in close relation to WQTS. 

file_name = 'PWSDEP_PT_WQTS.dbf'
pws = load_dbf_to_df(file_name)

pws.to_csv('../data/interim/pws_wqts.csv')

** PWSDEP_PT **

The DEP PWS datalayer has been compiled from several sources. The original DEP PWS point dataset contained only community water supplies (CWS) which were located and digitized from stable mylar overlays based on USGS topographic quadrangles. In 1993 locations for community and non-community sources were generated from global positioning system (GPS) data collected by the U.S. Environmental Protection Agency (EPA). The EPA GPS locations were merged with the existing DEP CWS point data to create the DEP PWS datalayer. From June 1996 to February 1997 the DEP GIS Program and Drinking Water Program (DWP) conducted an intensive update of the PWS datalayer. This update involved several methodologies including map interpolation and GPS field verification.

file_name = 'PWSDEP_PT.dbf'
pts = load_dbf_to_df(file_name)

pws['PWS_CLASS'].value_counts()

pts['TYPE'].value_counts()

#add public water supply data to each point
pts = pts.merge(pws, how='left', on='SOURCE_ID')

#drop proposed wells
pts = pts[pts['TYPE'] != 'PW']

Create dataframe of potential public water sources for schools, which could be:
* town-wise system
* town system specific to school

#PWS related to town water departments are all COM, with the following keywords in PWS_NAME
towns = pts['TOWN'].tolist()
for town in ['FOXBORO', 'CENTERVILLE', 'BARNSTABLE', 'COTUIT', 'ABINGTON', 'ROCKLAND']:
    towns.append(town)
muni_keywords = ['DEP', 'DPW', 'DIST', 'WORKS', 'DIV', 'COMM.', 'COMMISSION', 'SUPPLY', 'WD' 'SYS', 'WORKS', 'TOWN OF']

#mark PWS with muni keywords in PWS_NAME as potential school water supply (['SCH_SUPPLY']=1)
pts.loc[pts['PWS_NAME'].str.contains('|'.join(towns)) & pts['PWS_NAME'].str.contains('|'.join(muni_keywords)), 'SCH_SUPPLY'] = 1

#mark PWS with 'SCHOOL' in PWS_NAME as potential school water supply (['SCH_SUPPLY']=1)
pts.loc[pts['PWS_NAME'].str.contains('SCHOOL'), 'SCH_SUPPLY'] = 1

school_pts = pts[pts['SCH_SUPPLY'] == 1]

Create dataframe with one row for each PWS_ID/TOWN that includes:
* PWS_ID
* PWS_NAME
* TOWN
* PWS_CLASS
* COUNT_GW (# of groundwater intakes in system)
* COUNT_SW (# of surface water intakes in system)

#drop emergency surface water (ESW) pts and well field (WF)
school_pts = school_pts[school_pts['TYPE'] != 'ESW']
school_pts = school_pts[school_pts['TYPE'] != 'WF']

#keep only:
#active suppliers (PWS_STATUS = A)
#active Sources (S_STATUS = A)
#available Sources (S_AVAILABI = ACTIVE)
school_pts = school_pts[school_pts['PWS_STATUS'] == 'A']
school_pts = school_pts[school_pts['S_STATUS'] == 'A']
school_pts = school_pts[school_pts['S_AVAILABI'] == 'ACTIVE']

#find type for each intake based on SOURCE_ID
school_pts['INTAKE_TYPE'] = school_pts['SOURCE_ID'].apply(lambda x: x[-1])

sch_pws = school_pts[['PWS_ID', 'PWS_NAME', 'TOWN', 'PWS_CLASS']].drop_duplicates()

#calculate number of gw and sw intakes for each system
intakes = school_pts.groupby(['PWS_ID', 'INTAKE_TYPE']).count()[['PWS_NAME']]
intakes = intakes.unstack()
intakes.fillna(0.0, inplace=True)
intakes.reset_index(inplace=True)

sch_pws = sch_pws.merge(intakes, on='PWS_ID')

columns = 'PWS_ID', 'PWS_NAME', 'TOWN', 'PWS_CLASS', 'COUNT_GW', 'COUNT_SW'
sch_pws.columns = columns

Create GW ratio column

sch_pws['TOTAL_INTAKES'] = sch_pws[['COUNT_GW', 'COUNT_SW']].sum(axis=1)

sch_pws['GW_RATIO'] = sch_pws['COUNT_GW'].div(sch_pws['TOTAL_INTAKES'])

sch_pws.drop(['COUNT_GW', 'COUNT_SW'], axis=1, inplace=True)

sch_pws.head()

### Add MWRA systems


sch_pws['MWRA'] = 0

mwra = open_pickle('../data/interim/mwra.pkl')

sch_pws = pd.concat([sch_pws, mwra])

### Add # of service connections from PWS contact info csv

connections = pd.read_csv('../data/external/MassDEP/pwscont.csv',
                          usecols=['PWSID', '# SERVICE CONNECTIONS'], 
                          dtype='str')

connections.columns = ['PWS_ID', 'COUNT_CONN']

sch_pws = sch_pws.merge(connections, on='PWS_ID')

len(sch_pws)

### Add results and whether reported pb90 results from pb90 csv

pb90 = pd.read_csv('../data/external/MassDEP/pb90.csv', 
                   usecols=['PWSID', 'RESULT'],
                   dtype='str')
pb90.columns = ['PWS_ID', 'PB90_RESULT']

pb90.head()

sch_pws = sch_pws.merge(pb90, how='left', on='PWS_ID')

sch_pws.isnull().sum()/len(sch_pws)

sch_pws['PB90_RESULT'].fillna(-1, inplace=True)

sch_pws.loc[sch_pws['PB90_RESULT'] >= 0, 'PB_90'] = 1
sch_pws.loc[sch_pws['PB90_RESULT'] < 0, 'PB_90'] = 0

sch_pws.head()

#sch_pws.drop('RESULT', axis=1, inplace=True)

### Add whether responded to LCR survey from lcrsurvresp111616 csv

survey = pd.read_csv('../data/external/MassDEP/lcrsurvresp111616.csv', 
                   usecols=['Unnamed: 1', 'Unnamed: 2'],
                   dtype='str')
survey.columns = ['PWS_ID', 'SURVEY']

survey.drop(0, inplace=True)

survey['SURVEY'] = 1

sch_pws = sch_pws.merge(survey, how='left', on='PWS_ID')
sch_pws['SURVEY'].fillna(0, inplace=True)

sch_pws.isnull().sum()

make_pickle(sch_pws, '../data/interim/sch_pws.pkl')
