==> 03_mcnulty/notebooks/04_Make_features_target.ipynb <==
import pandas as pd
import numpy as np
import datetime as dt
import re
import string

from dbfread import DBF

import operator
import pickle

import matplotlib.pyplot as plt
%matplotlib inline

import seaborn as sns
sns.set_style('white')

from sklearn import preprocessing
from sklearn.model_selection import train_test_split
from mlxtend.preprocessing import standardize

#pickle objects
def make_pickle(obj, filename):
    with open(filename, 'wb') as picklefile:
        pickle.dump(obj, picklefile)

#open pickled objects
def open_pickle(filename):
    with open(filename, 'rb') as picklefile:
        return pickle.load(picklefile)

### Load and merge schools and PWS data

#load school df, remove punct, make upper case, keep relevant columns
schools = open_pickle('../data/interim/schools.pkl')
schools['NAME_y'].fillna(schools['NAME_x'], inplace=True)
schools['NAME_y'] = schools['NAME_y'].apply(lambda x: x.translate(None, string.punctuation).upper())
schools = schools[['NAME_y', 'ORGCODE', 'RESULT', 'TOWN_x', 'YEAR_BUILT', 'TOTAL_VAL', 'TYPE', 'EJ']]

len(schools)

#drop schools without building data
schools.dropna(inplace=True)

schools.isnull().sum()

len(schools)

#load pws df, remove punct, make upper case
pws = open_pickle('../data/interim/sch_pws.pkl')
pws['PWS_NAME'] = pws['PWS_NAME'].apply(lambda x: x.translate(None, string.punctuation).upper())

school_pws = pws[pws['PWS_NAME'].str.contains('SCHOOL')]
town_pws = pws[~pws['PWS_NAME'].str.contains('SCHOOL')]

#create substring for matching school names to pws name
schools['MATCH_NAME'] = schools['NAME_y'].apply(lambda x: re.findall('(\w+\s+\w+)', x))
schools.loc[schools['NAME_y'] == 'GOALS', 'MATCH_NAME'] = ['GOALS']
schools.loc[schools['NAME_y'] == 'MOPPETS', 'MATCH_NAME'] = ['MOPPETS']

schools['MATCH_NAME'] = schools['MATCH_NAME'].apply(lambda x: x[0])

school_pws['MATCH_NAME'] = school_pws['PWS_NAME'].apply(lambda x: re.findall('(\w+\s+\w+)', x)[0])

match_name_merge = schools.merge(school_pws, how='left', on='MATCH_NAME', indicator=True)

match_name_merge['_merge'].value_counts()

match_name = match_name_merge[match_name_merge['_merge'] == 'both']
schools = match_name_merge[match_name_merge['_merge'] == 'left_only']

drop_cols = list(school_pws.columns)
drop_cols.append('_merge')
schools.drop(labels=drop_cols, axis=1, inplace=True)

len(schools)

schools.columns

town_name = schools.merge(town_pws, how = 'left', left_on='TOWN_x', right_on='TOWN', indicator=True)

town_name['_merge'].value_counts()

town_name.dropna(inplace=True)
town_name.drop_duplicates(inplace=True)

#town_name[town_name[['NAME_y', 'TOWN']].duplicated(keep=False)].sort_values('NAME_y')[['NAME_y', 'TOWN_x']]
#town_name.drop_duplicates(['NAME_y', 'TOWN'], inplace=True)

sch_pws = pd.concat([town_name, match_name])

len(sch_pws)

sch_pws.drop_duplicates(subset=['NAME_y', 'TOWN_x'], inplace=True)

schools = sch_pws[['ORGCODE', 'NAME_y', 'TOWN', 'EJ', 'YEAR_BUILT', 'TOTAL_VAL', 'TYPE', 'RESULT', 'PWS_ID', 'PWS_NAME', 'PWS_CLASS', 'COUNT_CONN', 'TOTAL_INTAKES', 'GW_RATIO', 'SURVEY', 'PB90_RESULT', 'PB_90', 'MWRA']]

schools.rename(columns={'ORGCODE': 'SCH_ID',
               'NAME_y': 'SCH_NAME'},
              inplace=True)

### Create features

** Potential features related to parcels: **
* TOTAL_VAL: Current total assessed value for land and structures
* _LS__DATE: Last sale date formatted as YYYYMMDD - DROP FOR NOW_
* YEAR_BUILT: format YYYY

** Potential features related to schools: **
* TYPE: Identifies school type:
    * PUB - Public
    * PRI - Private
    * CHA - Charter
    * SPE - Special Education (Approved)
    * SPU - Special Education (Unapproved)
* _GRADES: Grade levels offered at the school - DROP FOR NOW_
* EJ

** Potential features related to PWS: **
* COUNT_CONN:
* TOTAL_INTAKES:
* GW_RATIO:
* SW_RATIO:
* SURVEY:
* PB_90:

** Target **
* RESULTS: max concentration (mg/L) of lead in samples collected from given school

schools.info()

for column in ['YEAR_BUILT', 'RESULT', 'TOTAL_VAL', 'COUNT_CONN', 'PB90_RESULT']:
    schools[column] = pd.to_numeric(schools[column])

#drop schools without year built data
schools = schools[schools['YEAR_BUILT'] != 0]
schools.reset_index(drop=True, inplace=True)

#fill empty val strings with NaN
schools['TOTAL_VAL'] = schools['TOTAL_VAL'].apply(lambda x: np.nan if not x else x)

### Clean features

dummies = ['SURVEY', 'PB_90', 'MWRA', 'EJ']
categoricals = ['TYPE', 'PWS_CLASS']
numericals = ['TOTAL_VAL', 'YEAR_BUILT', 'COUNT_CONN', 'TOTAL_INTAKES', 'GW_RATIO', 'PB90_RESULT']
target = ['RESULT']
features = dummies + categoricals + numericals + target
X = schools[features]

#fill missing total parcel value with median
#X.loc[:,'TOTAL_VAL'] = X.loc[:,'TOTAL_VAL'].fillna(X['TOTAL_VAL'].median())
X['TOTAL_VAL'].fillna(X['TOTAL_VAL'].median(), inplace=True)

#convert YEAR_BUILT to year since 1800 (earliest YEAR_BUILT)
#X.loc[:,'YEAR_SINCE_1800'] = X['YEAR_BUILT'].apply(lambda x: x - 1800)
X['YEAR_SINCE_1800'] = X['YEAR_BUILT'].apply(lambda x: x - 1800)
X = X.drop('YEAR_BUILT', axis=1)

#create dummy variables for categoricals
#all_dummies = pd.DataFrame(schools['RESULT'])
#for dummy in dummies:
#    all_dummies = all_dummies.merge(pd.DataFrame(X.pop(dummy)), left_index=True, right_index=True)

for category in categoricals:
    cat_dummies = pd.get_dummies(X[category], drop_first=True)
    X = X.merge(cat_dummies, left_index=True, right_index=True)
    X = X.drop(category, axis=1)

dummy_cols = {'PRI': 'SCH_PRI',
              'PUB': 'SCH_PUB', 
              'SPE': 'SCH_SPE', 
              'SPU': 'SCH_SPU', 
              'NC': 'PWS_NC', 
              'NTNC': 'PWS_NTNC'}
X.rename(columns=dummy_cols, inplace=True)

pb_al = 0.015

#for whether a result is greater than action level
X.loc[X['RESULT'] > pb_al, 'EXCEED'] = 1
X['EXCEED'].fillna(0, inplace=True)
X.drop('RESULT', axis=1, inplace=True)

X.head()

#use all PWS 
X.drop(['PB90_RESULT'], axis=1, inplace=True)

#use only PWS with pb90 results
#X = X[X['PB90_RESULT'] >= 0]
#X.drop(['PB_90'], axis=1, inplace=True)

y = X.pop('EXCEED')

len(X)

#train/test split, then cv on only training data, score on testing data
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=120, random_state=4444, stratify=y)
#X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=80, random_state=4444)

for df in [X_train, X_test, y_train, y_test]:
    df.reset_index(inplace=True, drop=True)

X_train.head()

s_cols = ['TOTAL_VAL', 'YEAR_SINCE_1800', 'COUNT_CONN', 'TOTAL_INTAKES']
X_train_s = X_train[s_cols]
X_test_s = X_test[s_cols]

#standardize numericals (train and test separately)
scaler = preprocessing.StandardScaler()
X_train_s = pd.DataFrame(scaler.fit_transform(X_train_s))
X_train_s.columns = s_cols

X_test_s = pd.DataFrame(scaler.transform(X_test_s))
X_test_s.columns = s_cols

X_test_s.head()

#standardize numericals (train and test separately)
X_train.drop(labels=s_cols, axis=1, inplace=True)
X_test.drop(labels=s_cols, axis=1, inplace=True)

X_train = X_train.merge(X_train_s, how='left', left_index=True, right_index=True)
X_test = X_test.merge(X_test_s, how='left', left_index=True, right_index=True)

make_pickle((X_train, X_test, y_train, y_test), '../data/processed/train_test.pkl')
